moves cdp related browser automation into its own dir
