+++ packages/server/lib/automation/helpers/evaluate_in_frame_context.ts
@@ -1,5 +1,5 @@
 import type { Protocol } from 'devtools-protocol'
-import type { SendDebuggerCommand } from '../../browsers/cdp_automation'
+import type { SendDebuggerCommand } from '../../browsers/cdp-protocol/cdp_automation'
 import Debug from 'debug'
 import { isError } from 'lodash'
 import type ProtocolMapping from 'devtools-protocol/types/protocol-mapping'
