+++ packages/server/lib/browsers/cdp-protocol/cdp-connection.ts
@@ -6,8 +6,8 @@ import type { CypressError } from '@packages/errors'
 import { debugCdpConnection, DebuggableCDPClient } from './debug-cdp-connection'
 import type { CdpEvent, CdpCommand } from './cdp_automation'
 import { CDPDisconnectedError, CDPTerminatedError, CDPAlreadyConnectedError } from './cri-errors'
-import { asyncRetry } from '../util/async_retry'
-import * as errors from '../errors'
+import { asyncRetry } from '../../util/async_retry'
+import * as errors from '../../errors'
 import type WebSocket from 'ws'
 
 const verboseDebugNs = 'cypress-verbose:server:browsers:cdp-connection'
