+++ packages/server/lib/browsers/chrome.ts
@@ -10,13 +10,13 @@ import { launch } from '@packages/launcher'
 
 import * as appData from '../util/app_data'
 import { fs } from '../util/fs'
-import { CdpAutomation, screencastOpts } from './cdp_automation'
+import { CdpAutomation, screencastOpts } from './cdp-protocol/cdp_automation'
 import * as protocol from './protocol'
 import utils from './utils'
 import * as errors from '../errors'
 import { BrowserCriClient } from './browser-cri-client'
 import type { Browser, BrowserInstance, GracefulShutdownOptions } from './types'
-import type { CriClient } from './cri-client'
+import type { CriClient } from './cdp-protocol/cri-client'
 import type { Automation } from '../automation'
 import memory from './memory'
