+++ packages/server/lib/browsers/webkit-automation.ts
@@ -1,7 +1,7 @@
 import Debug from 'debug'
 import type playwright from 'playwright-webkit'
 import type { Automation } from '../automation'
-import { normalizeResourceType } from './cdp_automation'
+import { normalizeResourceType } from './cdp-protocol/cdp_automation'
 import os from 'os'
 import type { RunModeVideoApi } from '@packages/types'
 import path from 'path'
